Improve ArC integration
- merge FullArchiveIndexBuildItem into BeanArchiveBuildItem
- fix the default annotation transformer - use special stereotype
instead of Dependent
