+++ independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/BeanProcessor.java
@@ -17,41 +17,25 @@
 package io.quarkus.arc.processor;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-import javax.enterprise.context.BeforeDestroyed;
-import javax.enterprise.context.Destroyed;
-import javax.enterprise.context.Initialized;
-import javax.enterprise.context.control.ActivateRequestContext;
-import javax.enterprise.inject.Any;
-import javax.enterprise.inject.Default;
-import javax.inject.Named;
-
 import org.jboss.jandex.AnnotationInstance;
-import org.jboss.jandex.ClassInfo;
 import org.jboss.jandex.CompositeIndex;
 import org.jboss.jandex.DotName;
-import org.jboss.jandex.Index;
 import org.jboss.jandex.IndexView;
 import org.jboss.jandex.Indexer;
-import org.jboss.jandex.Type;
 import org.jboss.logging.Logger;
 
-import io.quarkus.arc.ActivateRequestContextInterceptor;
 import io.quarkus.arc.processor.BuildExtension.BuildContext;
 import io.quarkus.arc.processor.BuildExtension.Key;
 import io.quarkus.arc.processor.DeploymentEnhancer.DeploymentContext;
@@ -70,7 +54,7 @@ public class BeanProcessor {
 
     static final String DEFAULT_NAME = "Default";
 
-    private static final Logger LOGGER = Logger.getLogger(BeanProcessor.class);
+    static final Logger LOGGER = Logger.getLogger(BeanProcessor.class);
 
     private final String name;
 
@@ -126,12 +110,12 @@ public class BeanProcessor {
 
                 @Override
                 public void addClass(String className) {
-                    index(indexer, className);
+                    BeanArchives.index(indexer, className);
                 }
 
                 @Override
                 public void addClass(Class<?> clazz) {
-                    index(indexer, clazz.getName());
+                    BeanArchives.index(indexer, clazz.getName());
                 }
 
                 @Override
@@ -234,31 +218,6 @@ public class BeanProcessor {
         }
         LOGGER.debugf("Generated %s resources in %s ms", resources.size(), System.currentTimeMillis() - start);
         return beanDeployment;
-    }
-
-    public static IndexView addBuiltinClasses(IndexView index) {
-        Indexer indexer = new Indexer();
-        // Add builtin interceptors and bindings
-        index(indexer, ActivateRequestContext.class.getName());
-        index(indexer, ActivateRequestContextInterceptor.class.getName());
-        // Add builtin qualifiers if needed
-        if (index.getClassByName(DotNames.ANY) == null) {
-            index(indexer, Default.class.getName());
-            index(indexer, Any.class.getName());
-            index(indexer, Named.class.getName());
-            index(indexer, Initialized.class.getName());
-            index(indexer, BeforeDestroyed.class.getName());
-            index(indexer, Destroyed.class.getName());
-        }
-        return new IndexWrapper(CompositeIndex.create(index, indexer.complete()));
-    }
-
-    private static void index(Indexer indexer, String className) {
-        try (InputStream stream = BeanProcessor.class.getClassLoader().getResourceAsStream(className.replace('.', '/') + ".class")) {
-            indexer.index(stream);
-        } catch (IOException e) {
-            throw new IllegalStateException("Failed to index: " + className, e);
-        }
     }
 
     public static class Builder {
@@ -416,153 +375,6 @@ public class BeanProcessor {
         }
         extensions.sort(BuildExtension::compare);
         return extensions;
-    }
-
-    /**
-     * This wrapper is used to index JDK classes on demand.
-     */
-    public static class IndexWrapper implements IndexView {
-
-        private final Map<DotName, ClassInfo> additionalClasses;
-
-        private final IndexView index;
-
-        public IndexWrapper(IndexView index) {
-            this.index = index;
-            this.additionalClasses = new ConcurrentHashMap<>();
-        }
-
-        @Override
-        public Collection<ClassInfo> getKnownClasses() {
-            return index.getKnownClasses();
-        }
-
-        @Override
-        public ClassInfo getClassByName(DotName className) {
-            ClassInfo classInfo = index.getClassByName(className);
-            if (classInfo == null) {
-                return additionalClasses.computeIfAbsent(className, name -> {
-                    LOGGER.debugf("Index: %s", className);
-                    Indexer indexer = new Indexer();
-                    index(indexer, className.toString());
-                    Index index = indexer.complete();
-                    return index.getClassByName(name);
-                });
-            }
-            return classInfo;
-        }
-
-        @Override
-        public Collection<ClassInfo> getKnownDirectSubclasses(DotName className) {
-            if (additionalClasses.isEmpty()) {
-                return index.getKnownDirectSubclasses(className);
-            }
-            Set<ClassInfo> directSubclasses = new HashSet<ClassInfo>(index.getKnownDirectSubclasses(className));
-            for (ClassInfo additional : additionalClasses.values()) {
-                if (className.equals(additional.superName())) {
-                    directSubclasses.add(additional);
-                }
-            }
-            return directSubclasses;
-        }
-
-        @Override
-        public Collection<ClassInfo> getAllKnownSubclasses(DotName className) {
-            if (additionalClasses.isEmpty()) {
-                return index.getAllKnownSubclasses(className);
-            }
-            final Set<ClassInfo> allKnown = new HashSet<ClassInfo>();
-            final Set<DotName> processedClasses = new HashSet<DotName>();
-            getAllKnownSubClasses(className, allKnown, processedClasses);
-            return allKnown;
-        }
-
-        @Override
-        public Collection<ClassInfo> getKnownDirectImplementors(DotName className) {
-            if (additionalClasses.isEmpty()) {
-                return index.getKnownDirectImplementors(className);
-            }
-            Set<ClassInfo> directImplementors = new HashSet<ClassInfo>(index.getKnownDirectImplementors(className));
-            for (ClassInfo additional : additionalClasses.values()) {
-                for (Type interfaceType : additional.interfaceTypes()) {
-                    if (className.equals(interfaceType.name())) {
-                        directImplementors.add(additional);
-                        break;
-                    }
-                }
-            }
-            return directImplementors;
-        }
-
-        @Override
-        public Collection<ClassInfo> getAllKnownImplementors(DotName interfaceName) {
-            if (additionalClasses.isEmpty()) {
-                return index.getAllKnownImplementors(interfaceName);
-            }
-            final Set<ClassInfo> allKnown = new HashSet<ClassInfo>();
-            final Set<DotName> subInterfacesToProcess = new HashSet<DotName>();
-            final Set<DotName> processedClasses = new HashSet<DotName>();
-            subInterfacesToProcess.add(interfaceName);
-            while (!subInterfacesToProcess.isEmpty()) {
-                final Iterator<DotName> toProcess = subInterfacesToProcess.iterator();
-                DotName name = toProcess.next();
-                toProcess.remove();
-                processedClasses.add(name);
-                getKnownImplementors(name, allKnown, subInterfacesToProcess, processedClasses);
-            }
-            return allKnown;
-        }
-
-        @Override
-        public Collection<AnnotationInstance> getAnnotations(DotName annotationName) {
-            return index.getAnnotations(annotationName);
-        }
-
-        private void getAllKnownSubClasses(DotName className, Set<ClassInfo> allKnown, Set<DotName> processedClasses) {
-            final Set<DotName> subClassesToProcess = new HashSet<DotName>();
-            subClassesToProcess.add(className);
-            while (!subClassesToProcess.isEmpty()) {
-                final Iterator<DotName> toProcess = subClassesToProcess.iterator();
-                DotName name = toProcess.next();
-                toProcess.remove();
-                processedClasses.add(name);
-                getAllKnownSubClasses(name, allKnown, subClassesToProcess, processedClasses);
-            }
-        }
-
-        private void getAllKnownSubClasses(DotName name, Set<ClassInfo> allKnown, Set<DotName> subClassesToProcess, Set<DotName> processedClasses) {
-            final Collection<ClassInfo> directSubclasses = getKnownDirectSubclasses(name);
-            if (directSubclasses != null) {
-                for (final ClassInfo clazz : directSubclasses) {
-                    final DotName className = clazz.name();
-                    if (!processedClasses.contains(className)) {
-                        allKnown.add(clazz);
-                        subClassesToProcess.add(className);
-                    }
-                }
-            }
-        }
-
-        private void getKnownImplementors(DotName name, Set<ClassInfo> allKnown, Set<DotName> subInterfacesToProcess, Set<DotName> processedClasses) {
-            final Collection<ClassInfo> list = getKnownDirectImplementors(name);
-            if (list != null) {
-                for (final ClassInfo clazz : list) {
-                    final DotName className = clazz.name();
-                    if (!processedClasses.contains(className)) {
-                        if (Modifier.isInterface(clazz.flags())) {
-                            subInterfacesToProcess.add(className);
-                        } else {
-                            if (!allKnown.contains(clazz)) {
-                                allKnown.add(clazz);
-                                processedClasses.add(className);
-                                getAllKnownSubClasses(className, allKnown, processedClasses);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
     }
 
     static class BuildContextImpl implements BuildContext {
